feat: Add password confirm and populate

+++ controllers/genreController.js
@@ -1,8 +1,9 @@
+require("dotenv").config();
 const { ResultWithContextImpl } = require("express-validator/lib/chain");
 const { body, validationResult, matchedData } = require("express-validator");
 const db = require("../db/genreQueries");
 
-const validateGenre = [
+const validateAdd = [
   body("name")
     .trim()
     .notEmpty()
@@ -12,6 +13,45 @@ const validateGenre = [
     .withMessage("Genre contains invalid characters.")
     .isLength({ max: 32 })
     .withMessage("Genre must not be longer than 32 characters."),
+];
+
+const validateEdit = [
+  body("name")
+    .trim()
+    .notEmpty()
+    .withMessage("Genre is required.")
+    .bail()
+    .matches(/^[\p{L}\d '&/-]+$/u)
+    .withMessage("Genre contains invalid characters.")
+    .isLength({ max: 32 })
+    .withMessage("Genre must not be longer than 32 characters."),
+  body("password")
+    .trim()
+    .notEmpty()
+    .withMessage("Password is required.")
+    .bail()
+    .custom((value) => {
+      if (value !== process.env.DB_DELETE) {
+        throw new Error("Invalid password");
+      }
+
+      return true;
+    }),
+];
+
+const validateDelete = [
+  body("password")
+    .trim()
+    .notEmpty()
+    .withMessage("Password is required.")
+    .bail()
+    .custom((value) => {
+      if (value !== process.env.DB_DELETE) {
+        throw new Error("Invalid password");
+      }
+
+      return true;
+    }),
 ];
 
 const getAllGenres = async (req, res) => {
@@ -24,7 +64,7 @@ const getNewGenre = async (req, res) => {
 };
 
 const postNewGenre = [
-  validateGenre,
+  validateAdd,
   async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -47,7 +87,7 @@ const getEditGenre = async (req, res) => {
 };
 
 const postEditGenre = [
-  validateGenre,
+  validateEdit,
   async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -64,10 +104,22 @@ const postEditGenre = [
   },
 ];
 
-const postDeleteGenre = async (req, res) => {
-  const genre = await db.deleteGenre(req.params.id);
-  res.redirect("/genres");
-};
+const postDeleteGenre = [
+  validateDelete,
+  async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      const genre = await db.readGenre(req.params.id);
+      return res.status(400).render("genres/genreEditForm", {
+        title: "Edit Genre",
+        genre: genre,
+        errors: errors.array(),
+      });
+    }
+    await db.deleteGenre(req.params.id);
+    res.redirect("/genres");
+  },
+];
 
 module.exports = {
   getAllGenres,
